Added service layer unit tests

+++ src/main/java/com/agatah/dogtalk/service/PetService.java
@@ -2,6 +2,7 @@ package com.agatah.dogtalk.service;
 
 import com.agatah.dogtalk.dto.PetProfileDto;
 import com.agatah.dogtalk.dto.mappers.PetProfileMapper;
+import com.agatah.dogtalk.exception.EntityNotFoundException;
 import com.agatah.dogtalk.model.PetOwnerProfile;
 import com.agatah.dogtalk.model.Pet;
 import com.agatah.dogtalk.model.Photo;
@@ -14,9 +15,9 @@ import java.util.Optional;
 @Service
 public class PetService {
 
-    private PetProfileRepository petProfileRepository;
-    private OwnerProfileRepository ownerProfileRepository;
-    private PhotoService photoService;
+    private final PetProfileRepository petProfileRepository;
+    private final OwnerProfileRepository ownerProfileRepository;
+    private final PhotoService photoService;
 
     public PetService(PetProfileRepository petProfileRepository, OwnerProfileRepository ownerProfileRepository,
                       PhotoService photoService){
@@ -35,30 +36,36 @@ public class PetService {
                     .setBreed(petProfile.getBreed());
             return PetProfileMapper.toPetProfileDto(petProfileRepository.save(pet));
         } else {
-            return null;
+            throw new EntityNotFoundException(PetOwnerProfile.class, ownerId);
         }
     }
 
     public void deletePetProfile(Long petId){
-        petProfileRepository.deleteById(petId);
+        if(petProfileRepository.existsById(petId)){
+            petProfileRepository.deleteById(petId);
+        }
     }
 
     public PetProfileDto updatePetProfile(PetProfileDto petProfileDto){
         Optional<Pet> petProfileOpt = petProfileRepository.findById(petProfileDto.getPetId());
         if(petProfileOpt.isPresent()){
             Pet dbPet = petProfileOpt.get();
-            dbPet.setName(petProfileDto.getPetName());
+            dbPet.setName(petProfileDto.getPetName())
+                    .setAge(petProfileDto.getAge())
+                    .setBreed(petProfileDto.getBreed());
             return PetProfileMapper.toPetProfileDto(petProfileRepository.save(dbPet));
         } else {
-            return null;
+            throw new EntityNotFoundException(Pet.class, petProfileDto.getPetId());
         }
     }
 
     public PetProfileDto addPhoto(Long petId, Photo photo){
         Optional<Pet> petOpt = petProfileRepository.findById(petId);
-        if(petOpt.isPresent()){
-            PetProfileMapper.toPetProfileDto(petProfileRepository.save(petOpt.get().setPhoto(photoService.uploadPhoto(photo))));
-        }
-        return null;
+        return petOpt
+                .map(pet -> pet.setPhoto(photoService.uploadPhoto(photo)))
+                .map(petProfileRepository::save)
+                .map(PetProfileMapper::toPetProfileDto)
+                .orElseThrow(() -> new EntityNotFoundException(Pet.class, petId));
+
     }
 }
